passport-photo-gemini/requirements.txt
streamlit
opencv-python-headless
numpy
pillow
rembg
onnxruntime
pyyaml
typeguard